feat: Typo fixed

+++ java/src/ru/koshakmine/icstd/ui/elements/CloseButtonElement.java
@@ -1,11 +1,11 @@
 package ru.koshakmine.icstd.ui.elements;
 
 public class CloseButtonElement extends ButtonElement {
-    public CloseButtonElement(int x, int y, String bitmap, float scale){
+    public CloseButtonElement(float x, float y, String bitmap, float scale){
         super("closeButton", x, y, bitmap, scale);
     }
 
-    public CloseButtonElement(int x, int y, String bitmap){
+    public CloseButtonElement(float x, float y, String bitmap){
         super("closeButton", x, y, bitmap, 1f);
     }
 }
